refactor(admin): split form-editor metabox renderer into per-metabox classes (S3 of #141)
Pure code movement — no behavior changes. The single
FormEditorMetaboxRenderer (1245 LOC) becomes a thin facade (99 LOC)
that delegates each metabox to its own class.

Nine new per-metabox classes (one render(WP_Post $post) method each):

  - FormEditorShortcodeMetabox             52 LOC
  - FormEditorLayoutMetabox                98 LOC
  - FormEditorBuilderMetabox              201 LOC (incl. render_field_row helper)
  - FormEditorRestrictionMetabox          152 LOC
  - FormEditorEmailMetabox                 89 LOC
  - FormEditorGeofenceMetabox             329 LOC (incl. inline toggleGeoSource JS)
  - FormEditorQuizMetabox                  87 LOC
  - FormEditorPublicCsvDownloadMetabox    300 LOC
  - FormEditorDeviceLimitMetabox          151 LOC

The facade `FormEditorMetaboxRenderer` keeps its public API intact
(render_box_*, render_shortcode_metabox, render_field_row) so existing
call sites in `FormEditor::add_custom_metaboxes()` and the unit test
work unchanged. Each public method is now a one-line forwarder.

PSR-4 autoloader resolves all 9 new class names from kebab-case
filenames automatically — no loader/composer changes needed.

Verification:
  - php -l: clean on all 10 files.
  - vendor/bin/phpunit --filter FormEditorMetaboxRenderer|FormEditorSaveHandler: 29/29 OK.
  - vendor/bin/phpstan analyse includes/admin/ --level=8: no errors.
