+++ includes/admin/class-ffc-form-editor-metabox-renderer.php
@@ -1,9 +1,14 @@
 <?php
 /**
- * Form Editor Metabox Renderer
+ * Form Editor Metabox Renderer (facade)
  *
  * Handles rendering of all metaboxes for the Form Editor.
  * Extracted from FFC_Form_Editor class to follow Single Responsibility Principle.
+ *
+ * Since S3 of the god-object refactor (issue #141), this class is a thin
+ * facade that delegates each metabox to its own per-metabox renderer class.
+ * The public API is preserved so that existing call sites in
+ * {@see FormEditor::add_custom_metaboxes()} continue to work unchanged.
  *
  * @since   3.1.1
  * @package FreeFormCertificate\Admin
@@ -13,7 +18,6 @@ declare(strict_types=1);
 
 namespace FreeFormCertificate\Admin;
 
-use FreeFormCertificate\Security\GeofenceLocationRegistry;
 use WP_Post;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -27,1219 +31,72 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class FormEditorMetaboxRenderer {
 
+	private FormEditorShortcodeMetabox $shortcode;
+	private FormEditorLayoutMetabox $layout;
+	private FormEditorBuilderMetabox $builder;
+	private FormEditorRestrictionMetabox $restriction;
+	private FormEditorEmailMetabox $email;
+	private FormEditorGeofenceMetabox $geofence;
+	private FormEditorQuizMetabox $quiz;
+	private FormEditorPublicCsvDownloadMetabox $public_csv_download;
+	private FormEditorDeviceLimitMetabox $device_limit;
+
 	/**
-	 * Render the shortcode sidebar metabox
-	 *
-	 * @param WP_Post $post The post object.
+	 * Wire up the per-metabox renderers.
 	 */
+	public function __construct() {
+		$this->shortcode           = new FormEditorShortcodeMetabox();
+		$this->layout              = new FormEditorLayoutMetabox();
+		$this->builder             = new FormEditorBuilderMetabox();
+		$this->restriction         = new FormEditorRestrictionMetabox();
+		$this->email               = new FormEditorEmailMetabox();
+		$this->geofence            = new FormEditorGeofenceMetabox();
+		$this->quiz                = new FormEditorQuizMetabox();
+		$this->public_csv_download = new FormEditorPublicCsvDownloadMetabox();
+		$this->device_limit        = new FormEditorDeviceLimitMetabox();
+	}
+
 	public function render_shortcode_metabox( WP_Post $post ): void {
-		?>
-		<div class="ffc-shortcode-box">
-			<p><strong><?php esc_html_e( 'Copy this Shortcode:', 'ffcertificate' ); ?></strong></p>
-			<code class="ffc-shortcode-display">
-				[ffc_form id="<?php echo esc_attr( (string) $post->ID ); ?>"]
-			</code>
-			<p class="description">
-				<?php esc_html_e( 'Paste this code into any Page or Post to display the form.', 'ffcertificate' ); ?>
-			</p>
-		</div>
-		<hr>
-		<p><strong><?php esc_html_e( 'Tips:', 'ffcertificate' ); ?></strong></p>
-		<ul class="ffc-tips-list">
-			<li><?php echo wp_kses_post( __( 'Use <b>{{field_name}}</b> in the PDF Layout to insert user data.', 'ffcertificate' ) ); ?></li>
-			<li><?php esc_html_e( 'Common variables include {{auth_code}}, {{submission_date}}, and {{ticket}}.', 'ffcertificate' ); ?></li>
-		</ul>
-		<?php
+		$this->shortcode->render( $post );
 	}
 
-	/**
-	 * Section 1: Certificate Layout Editor
-	 *
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_layout( WP_Post $post ): void {
-		$config   = get_post_meta( $post->ID, '_ffc_form_config', true );
-		$layout   = isset( $config['pdf_layout'] ) ? $config['pdf_layout'] : '';
-		$bg_image = isset( $config['bg_image'] ) ? $config['bg_image'] : '';
-
-		$templates_dir = FFC_PLUGIN_DIR . 'html/';
-		$templates     = glob( $templates_dir . '*.html' );
-
-		wp_nonce_field( 'ffc_save_form_data', 'ffc_form_nonce' );
-		?>
-		<table class="form-table">
-			<tr>
-				<th><label><?php esc_html_e( 'Actions', 'ffcertificate' ); ?></label></th>
-				<td>
-					<div class="ffc-admin-flex-row ffc-flex-wrap">
-						<div class="ffc-action-group">
-							<input type="file" id="ffc_import_html_file" accept=".html,.txt" class="ffc-hidden">
-							<button type="button" class="button" id="ffc_btn_import_html">
-								<?php esc_html_e( 'Import HTML', 'ffcertificate' ); ?>
-							</button>
-							<button type="button" class="button" id="ffc_btn_media_lib">
-								<?php esc_html_e( 'Background Image', 'ffcertificate' ); ?>
-							</button>
-							<button type="button" class="button button-primary" id="ffc_btn_preview">
-								<span class="dashicons dashicons-visibility" style="vertical-align:text-bottom;margin-right:2px;"></span>
-								<?php esc_html_e( 'Preview', 'ffcertificate' ); ?>
-							</button>
-						</div>
-
-						<?php if ( $templates ) : ?>
-						<div class="ffc-template-loader">
-							<select id="ffc_template_select">
-								<option value=""><?php esc_html_e( 'Select Server Template...', 'ffcertificate' ); ?></option>
-								<?php
-								foreach ( $templates as $tpl ) :
-									$filename = basename( $tpl );
-									?>
-									<option value="<?php echo esc_attr( $filename ); ?>"><?php echo esc_html( $filename ); ?></option>
-								<?php endforeach; ?>
-							</select>
-							<button type="button" id="ffc_load_template_btn" class="button"><?php esc_html_e( 'Load', 'ffcertificate' ); ?></button>
-						</div>
-						<?php endif; ?>
-					</div>
-				</td>
-			</tr>
-			<tr>
-				<td colspan="2">
-					<label class="ffc-block-label"><strong><?php esc_html_e( 'Certificate HTML Editor', 'ffcertificate' ); ?></strong></label>
-					<div class="ffc-code-editor-wrapper">
-						<textarea name="ffc_config[pdf_layout]" id="ffc_pdf_layout" class="ffc-w100" rows="12"><?php echo esc_textarea( $layout ); ?></textarea>
-					</div>
-					<p class="description">
-						<?php esc_html_e( 'Mandatory Tags:', 'ffcertificate' ); ?> <code>{{auth_code}}</code>, <code>{{name}}</code>, <code>{{cpf_rf}}</code>.
-					</p>
-
-					<div class="ffc-input-group ffc-mt15">
-						<label class="ffc-block-label"><strong><?php esc_html_e( 'Background Image URL:', 'ffcertificate' ); ?></strong></label>
-						<input type="text" name="ffc_config[bg_image]" id="ffc_bg_image_input" value="<?php echo esc_url( $bg_image ); ?>" class="ffc-w100">
-					</div>
-				</td>
-			</tr>
-		</table>
-		<?php
+		$this->layout->render( $post );
 	}
 
-	/**
-	 * Section 2: Form Builder (Fields)
-	 *
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_builder( WP_Post $post ): void {
-		$fields = get_post_meta( $post->ID, '_ffc_form_fields', true );
-
-		// Default fields for brand new forms.
-		if ( empty( $fields ) && 'auto-draft' === $post->post_status ) {
-			$fields = array(
-				array(
-					'type'     => 'text',
-					'label'    => __( 'Full Name', 'ffcertificate' ),
-					'name'     => 'name',
-					'required' => '1',
-					'options'  => '',
-				),
-				array(
-					'type'     => 'email',
-					'label'    => __( 'Email', 'ffcertificate' ),
-					'name'     => 'email',
-					'required' => '1',
-					'options'  => '',
-				),
-				array(
-					'type'     => 'text',
-					'label'    => __( 'CPF / ID', 'ffcertificate' ),
-					'name'     => 'cpf_rf',
-					'required' => '1',
-					'options'  => '',
-				),
-			);
-		}
-		?>
-		<div id="ffc-fields-container">
-			<?php
-			if ( ! empty( $fields ) && is_array( $fields ) ) {
-				foreach ( $fields as $index => $field ) {
-					$this->render_field_row( $index, $field );
-				}
-			}
-			?>
-		</div>
-		<div>
-		<p class="description">
-		<?php esc_html_e( 'Minimal fields (Tag):', 'ffcertificate' ); ?> <code>name</code>, <code>email</code>, <code>cpf_rf</code>.
-		</p>
-		</div>
-		<div class="ffc-builder-actions ffc-mt20">
-			<button type="button" class="button button-primary ffc-add-field">
-				<span class="dashicons dashicons-plus-alt"></span>
-				<?php esc_html_e( 'Add New Field', 'ffcertificate' ); ?>
-			</button>
-		</div>
-
-		<div class="ffc-field-template ffc-hidden">
-			<?php $this->render_field_row( 'TEMPLATE', array() ); ?>
-		</div>
-		<?php
+		$this->builder->render( $post );
 	}
 
-	/**
-	 * Section 3: Restrictions and Tickets
-	 *
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_restriction( WP_Post $post ): void {
-		$config = get_post_meta( $post->ID, '_ffc_form_config', true );
-
-		// Get restrictions (new structure).
-		$restrictions     = isset( $config['restrictions'] ) ? $config['restrictions'] : array();
-		$password_active  = ! empty( $restrictions['password'] ) && '1' === $restrictions['password'];
-		$allowlist_active = ! empty( $restrictions['allowlist'] ) && '1' === $restrictions['allowlist'];
-		$denylist_active  = ! empty( $restrictions['denylist'] ) && '1' === $restrictions['denylist'];
-		$ticket_active    = ! empty( $restrictions['ticket'] ) && '1' === $restrictions['ticket'];
-
-		// Legacy fields.
-		$vcode     = isset( $config['validation_code'] ) ? $config['validation_code'] : '';
-		$allow     = isset( $config['allowed_users_list'] ) ? $config['allowed_users_list'] : '';
-		$deny      = isset( $config['denied_users_list'] ) ? $config['denied_users_list'] : '';
-		$gen_codes = isset( $config['generated_codes_list'] ) ? $config['generated_codes_list'] : '';
-		?>
-		<table class="form-table">
-			<tr>
-				<th><label><?php esc_html_e( 'Form Restrictions', 'ffcertificate' ); ?></label></th>
-				<td>
-					<p class="description ffc-mb-15">
-						<?php esc_html_e( 'Select which restrictions to apply (can combine multiple):', 'ffcertificate' ); ?>
-					</p>
-
-					<label class="ffc-restriction-label">
-						<input type="checkbox"
-								name="ffc_config[restrictions][password]"
-								value="1"
-								id="ffc_restriction_password"
-								<?php checked( $password_active, true ); ?>>
-						<strong><?php esc_html_e( 'Single Password', 'ffcertificate' ); ?></strong>
-						<span class="description"> — <?php esc_html_e( 'Shared password for all users', 'ffcertificate' ); ?></span>
-					</label>
-
-					<label class="ffc-restriction-label">
-						<input type="checkbox"
-								name="ffc_config[restrictions][allowlist]"
-								value="1"
-								id="ffc_restriction_allowlist"
-								<?php checked( $allowlist_active, true ); ?>>
-						<strong><?php esc_html_e( 'Allowlist (CPF/RF)', 'ffcertificate' ); ?></strong>
-						<span class="description"> — <?php esc_html_e( 'Only approved CPF/RF can submit', 'ffcertificate' ); ?></span>
-					</label>
-
-					<label class="ffc-restriction-label">
-						<input type="checkbox"
-								name="ffc_config[restrictions][denylist]"
-								value="1"
-								id="ffc_restriction_denylist"
-								<?php checked( $denylist_active, true ); ?>>
-						<strong><?php esc_html_e( 'Denylist (CPF/RF)', 'ffcertificate' ); ?></strong>
-						<span class="description"> — <?php esc_html_e( 'Blocked CPF/RF cannot submit', 'ffcertificate' ); ?></span>
-					</label>
-
-					<label class="ffc-restriction-label">
-						<input type="checkbox"
-								name="ffc_config[restrictions][ticket]"
-								value="1"
-								id="ffc_restriction_ticket"
-								<?php checked( $ticket_active, true ); ?>>
-						<strong><?php esc_html_e( 'Ticket (Unique Codes)', 'ffcertificate' ); ?></strong>
-						<span class="description"> — <?php esc_html_e( 'Requires valid ticket (consumed after use)', 'ffcertificate' ); ?></span>
-					</label>
-
-					<p class="description ffc-mt-15">
-						<em><?php esc_html_e( 'Note: If no restriction is selected, form is Open (no restrictions).', 'ffcertificate' ); ?></em>
-					</p>
-				</td>
-			</tr>
-
-			<tr id="ffc_password_field" class="ffc-conditional-field<?php echo esc_attr( $password_active ? ' active' : '' ); ?>">
-				<th><label><?php esc_html_e( 'Password Value', 'ffcertificate' ); ?></label></th>
-				<td>
-					<input type="text"
-							name="ffc_config[validation_code]"
-							value="<?php echo esc_attr( $vcode ); ?>"
-							class="regular-text"
-							placeholder="<?php esc_attr_e( 'Ex: PASS2025', 'ffcertificate' ); ?>">
-					<p class="description"><?php esc_html_e( 'This password will be required from all users.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-
-			<tr id="ffc_allowlist_field" class="ffc-conditional-field<?php echo esc_attr( $allowlist_active ? ' active' : '' ); ?>">
-				<th><label><?php esc_html_e( 'Allowlist (CPFs / IDs)', 'ffcertificate' ); ?></label></th>
-				<td>
-					<textarea name="ffc_config[allowed_users_list]"
-								class="ffc-textarea-mono ffc-h120 ffc-w100"
-								placeholder="<?php esc_attr_e( 'One per line...', 'ffcertificate' ); ?>"><?php echo esc_textarea( $allow ); ?></textarea>
-					<p class="description"><?php esc_html_e( 'Accepts formats: 12345678900 or 123.456.789-00', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-
-			<tr id="ffc_denylist_field" class="ffc-conditional-field<?php echo esc_attr( $denylist_active ? ' active' : '' ); ?>">
-				<th><label><?php esc_html_e( 'Denylist (Blocked)', 'ffcertificate' ); ?></label></th>
-				<td>
-					<textarea name="ffc_config[denied_users_list]"
-								class="ffc-textarea-mono ffc-h80 ffc-w100"
-								placeholder="<?php esc_attr_e( 'Banned users...', 'ffcertificate' ); ?>"><?php echo esc_textarea( $deny ); ?></textarea>
-					<p class="description"><?php esc_html_e( 'Has priority over Allowlist. Accepts same formats.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-
-			<tr id="ffc_ticket_field" class="ffc-highlight-row ffc-conditional-field<?php echo esc_attr( $ticket_active ? ' active' : '' ); ?>">
-				<th><label class="ffc-label-accent"><?php esc_html_e( 'Ticket Generator', 'ffcertificate' ); ?></label></th>
-				<td>
-					<div class="ffc-admin-flex-row ffc-mb5">
-						<input type="number" id="ffc_qty_codes" value="10" min="1" max="500" class="ffc-input-small">
-						<button type="button" class="button button-secondary" id="ffc_btn_generate_codes"><?php esc_html_e( 'Generate Tickets', 'ffcertificate' ); ?></button>
-						<span id="ffc_gen_status" class="ffc-gen-status"></span>
-					</div>
-					<textarea name="ffc_config[generated_codes_list]"
-								id="ffc_generated_list"
-								class="ffc-textarea-mono ffc-h120 ffc-w100"><?php echo esc_textarea( $gen_codes ); ?></textarea>
-					<p class="description"><?php esc_html_e( 'Tickets are consumed (removed) after successful use.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-		</table>
-		<?php
+		$this->restriction->render( $post );
 	}
 
-	/**
-	 * Section 4: Email Settings
-	 *
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_email( WP_Post $post ): void {
-		$config     = get_post_meta( $post->ID, '_ffc_form_config', true );
-		$send_email = isset( $config['send_user_email'] ) ? $config['send_user_email'] : '0';
-		$subject    = isset( $config['email_subject'] ) ? $config['email_subject'] : __( 'Your Certificate', 'ffcertificate' );
-		$body       = isset( $config['email_body'] ) ? $config['email_body'] : '';
-		?>
-		<table class="form-table">
-			<tr>
-				<th><label><?php esc_html_e( 'Send Email to User?', 'ffcertificate' ); ?></label></th>
-				<td>
-					<select name="ffc_config[send_user_email]" class="ffc-select-full">
-						<option value="0" <?php selected( $send_email, '0' ); ?>><?php esc_html_e( 'No', 'ffcertificate' ); ?></option>
-						<option value="1" <?php selected( $send_email, '1' ); ?>><?php esc_html_e( 'Yes', 'ffcertificate' ); ?></option>
-					</select>
-				</td>
-			</tr>
-			<tr>
-				<th><label><?php esc_html_e( 'Subject', 'ffcertificate' ); ?></label></th>
-				<td><input type="text" name="ffc_config[email_subject]" value="<?php echo esc_attr( $subject ); ?>" class="ffc-w100"></td>
-			</tr>
-			<tr>
-				<th><label for="ffc_email_body"><?php esc_html_e( 'Email Body (HTML)', 'ffcertificate' ); ?></label></th>
-				<td>
-					<?php
-					wp_editor(
-						(string) $body,
-						'ffc_email_body',
-						array(
-							'textarea_name' => 'ffc_config[email_body]',
-							'textarea_rows' => 10,
-							'media_buttons' => false,
-							'teeny'         => true,
-							'tinymce'       => array(
-								'toolbar1' => 'bold,italic,underline,bullist,numlist,link,unlink,undo,redo',
-								'toolbar2' => '',
-							),
-							'quicktags'     => array( 'buttons' => 'strong,em,link,ul,ol,li,close' ),
-						)
-					);
-					?>
-					<p class="description">
-						<?php esc_html_e( 'Placeholders such as {{auth_code}} and {{name}} are preserved automatically.', 'ffcertificate' ); ?>
-					</p>
-				</td>
-			</tr>
-			<tr>
-				<th></th>
-				<td>
-				<p class="description ffc-mt-15">
-				<em><?php esc_html_e( 'Note: When this option is enabled, the email will only be sent when the user submits the form. This will add them to a waiting list and emails will be sent progressively.', 'ffcertificate' ); ?></em>
-				</p>
-				</td>
-			</tr>
-		</table>
-		<?php
+		$this->email->render( $post );
 	}
 
-	/**
-	 * Render Geofence & DateTime Restrictions Meta Box
-	 *
-	 * @since 3.0.0
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_geofence( WP_Post $post ): void {
-		$config = get_post_meta( $post->ID, '_ffc_geofence_config', true );
-		if ( ! is_array( $config ) ) {
-			$config = array();
-		}
-
-		// Defaults.
-		$datetime_enabled   = ( $config['datetime_enabled'] ?? '0' ) === '1' ? '1' : '0';
-		$date_start         = $config['date_start'] ?? '';
-		$date_end           = $config['date_end'] ?? '';
-		$time_start         = $config['time_start'] ?? '';
-		$time_end           = $config['time_end'] ?? '';
-		$time_mode          = $config['time_mode'] ?? 'daily'; // 'daily' or 'span'
-		$datetime_hide_mode = $config['datetime_hide_mode'] ?? 'message';
-		$msg_datetime       = $config['msg_datetime'] ?? __( 'This form is not available at this time.', 'ffcertificate' );
-
-		$geo_enabled              = ( $config['geo_enabled'] ?? '0' ) === '1' ? '1' : '0';
-		$geo_gps_enabled          = ( $config['geo_gps_enabled'] ?? '0' ) === '1' ? '1' : '0';
-		$geo_ip_enabled           = ( $config['geo_ip_enabled'] ?? '0' ) === '1' ? '1' : '0';
-		$geo_areas                = $config['geo_areas'] ?? '';
-		$geo_area_source          = $config['geo_area_source'] ?? 'custom';
-		$geo_area_location_ids    = $config['geo_area_location_ids'] ?? array();
-		$geo_ip_areas_permissive  = ( $config['geo_ip_areas_permissive'] ?? '0' ) === '1' ? '1' : '0';
-		$geo_ip_areas             = $config['geo_ip_areas'] ?? '';
-		$geo_ip_area_source       = $config['geo_ip_area_source'] ?? 'custom';
-		$geo_ip_area_location_ids = $config['geo_ip_area_location_ids'] ?? array();
-
-		if ( 'auto-draft' === $post->post_status ) {
-			$default_gps = GeofenceLocationRegistry::get_default_gps();
-			if ( $default_gps ) {
-				$geo_area_source       = 'locations';
-				$geo_area_location_ids = array( $default_gps['id'] );
-			}
-			$default_ip = GeofenceLocationRegistry::get_default_ip();
-			if ( $default_ip ) {
-				$geo_ip_area_source       = 'locations';
-				$geo_ip_area_location_ids = array( $default_ip['id'] );
-			}
-		}
-
-		$all_locations    = GeofenceLocationRegistry::get_all();
-		$geo_gps_ip_logic = $config['geo_gps_ip_logic'] ?? 'or';
-		$geo_hide_mode    = $config['geo_hide_mode'] ?? 'message';
-		$msg_geo_blocked  = $config['msg_geo_blocked'] ?? __( 'This form is not available in your location.', 'ffcertificate' );
-		$msg_geo_error    = $config['msg_geo_error'] ?? __( 'Unable to determine your location. Please enable location services.', 'ffcertificate' );
-		?>
-
-		<div class="ffc-geofence-container">
-			<!-- Sentinel: ensures ffc_geofence is always in POST even when all fields are disabled -->
-			<input type="hidden" name="ffc_geofence[_save]" value="1">
-			<!-- Tab Navigation -->
-			<div class="ffc-geofence-tabs">
-				<button type="button" class="ffc-geo-tab-btn active ffc-icon-calendar" data-tab="datetime">
-					<?php esc_html_e( 'Date & Time', 'ffcertificate' ); ?>
-				</button>
-				<button type="button" class="ffc-geo-tab-btn ffc-icon-globe" data-tab="geolocation">
-					<?php esc_html_e( 'Geolocation', 'ffcertificate' ); ?>
-				</button>
-			</div>
-
-			<!-- Tab: Date & Time -->
-			<div class="ffc-geo-tab-content active" id="ffc-tab-datetime">
-				<table class="form-table">
-					<tr>
-						<th><label><?php esc_html_e( 'Enable Date/Time Restrictions', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label>
-								<input type="checkbox" name="ffc_geofence[datetime_enabled]" value="1" <?php checked( $datetime_enabled, '1' ); ?>>
-								<?php esc_html_e( 'Restrict form access by date and time', 'ffcertificate' ); ?>
-							</label>
-							<p class="description"><?php esc_html_e( 'Control when users can access this form based on date range and daily hours.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Date Range', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label><?php esc_html_e( 'Start:', 'ffcertificate' ); ?> <input type="date" name="ffc_geofence[date_start]" value="<?php echo esc_attr( $date_start ); ?>"></label>
-							&nbsp;&nbsp;
-							<label><?php esc_html_e( 'End:', 'ffcertificate' ); ?> <input type="date" name="ffc_geofence[date_end]" value="<?php echo esc_attr( $date_end ); ?>"></label>
-							<p class="description"><?php esc_html_e( 'Leave empty for no date restriction. Format: YYYY-MM-DD', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Time Range', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label><?php esc_html_e( 'From:', 'ffcertificate' ); ?> <input type="time" name="ffc_geofence[time_start]" value="<?php echo esc_attr( $time_start ); ?>"></label>
-							&nbsp;&nbsp;
-							<label><?php esc_html_e( 'To:', 'ffcertificate' ); ?> <input type="time" name="ffc_geofence[time_end]" value="<?php echo esc_attr( $time_end ); ?>"></label>
-							<p class="description"><?php esc_html_e( 'Leave empty for 24/7 access. Default: 00:00 to 23:59', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr id="ffc-time-mode-row" class="ffc-conditional-field<?php echo esc_attr( ( ! empty( $date_start ) && ! empty( $date_end ) && $date_start !== $date_end ) ? ' active' : '' ); ?>">
-						<th><label><?php esc_html_e( 'Time Behavior', 'ffcertificate' ); ?></label></th>
-						<td>
-							<fieldset>
-								<label>
-									<input type="radio" name="ffc_geofence[time_mode]" value="span" <?php checked( $time_mode, 'span' ); ?>>
-									<strong><?php esc_html_e( 'Time spans across dates', 'ffcertificate' ); ?></strong>
-								</label>
-								<p class="description ffc-time-description">
-									<?php esc_html_e( 'Start time applies to start date, end time applies to end date. Form is open continuously between those timestamps.', 'ffcertificate' ); ?><br>
-									<?php esc_html_e( 'Example: Start 01/01 12:00 + End 10/01 23:00 = Open from 12:00 on Jan 1st until 23:00 on Jan 10th', 'ffcertificate' ); ?>
-								</p>
-
-								<label>
-									<input type="radio" name="ffc_geofence[time_mode]" value="daily" <?php checked( $time_mode, 'daily' ); ?>>
-									<strong><?php esc_html_e( 'Time applies to each day individually', 'ffcertificate' ); ?></strong>
-								</label>
-								<p class="description ffc-time-description">
-									<?php esc_html_e( 'Time range applies to every day in the date range. Form respects daily hours.', 'ffcertificate' ); ?><br>
-									<?php esc_html_e( 'Example: Start 01/01 + End 10/01 + Time 12:00-23:00 = Open 12:00-23:00 every day from Jan 1-10', 'ffcertificate' ); ?>
-								</p>
-							</fieldset>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Display Mode', 'ffcertificate' ); ?></label></th>
-						<td>
-							<select name="ffc_geofence[datetime_hide_mode]">
-								<option value="message" <?php selected( $datetime_hide_mode, 'message' ); ?>><?php esc_html_e( 'Show blocked message (Recommended)', 'ffcertificate' ); ?></option>
-								<option value="title_message" <?php selected( $datetime_hide_mode, 'title_message' ); ?>><?php esc_html_e( 'Show title + description + message', 'ffcertificate' ); ?></option>
-								<option value="hide" <?php selected( $datetime_hide_mode, 'hide' ); ?>><?php esc_html_e( 'Hide form completely', 'ffcertificate' ); ?></option>
-							</select>
-							<p class="description"><?php esc_html_e( 'How to display the form when date/time is invalid.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Blocked Message', 'ffcertificate' ); ?></label></th>
-						<td>
-							<textarea name="ffc_geofence[msg_datetime]" rows="3" class="ffc-w100"><?php echo esc_textarea( $msg_datetime ); ?></textarea>
-							<p class="description"><?php esc_html_e( 'Message shown when form is accessed outside allowed date/time.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-				</table>
-			</div>
-
-			<!-- Tab: Geolocation -->
-			<div class="ffc-geo-tab-content" id="ffc-tab-geolocation">
-				<table class="form-table">
-					<tr>
-						<th><label><?php esc_html_e( 'Enable Geolocation', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label>
-								<input type="checkbox" name="ffc_geofence[geo_enabled]" value="1" <?php checked( $geo_enabled, '1' ); ?>>
-								<?php esc_html_e( 'Restrict form access by geographic location', 'ffcertificate' ); ?>
-							</label>
-							<p class="description"><?php esc_html_e( 'Limit form access to users within specific geographic areas.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Validation Methods', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label>
-								<input type="checkbox" name="ffc_geofence[geo_gps_enabled]" value="1" <?php checked( $geo_gps_enabled, '1' ); ?>>
-								<?php esc_html_e( 'GPS (Browser geolocation)', 'ffcertificate' ); ?>
-							</label><br>
-							<label>
-								<input type="checkbox" name="ffc_geofence[geo_ip_enabled]" value="1" <?php checked( $geo_ip_enabled, '1' ); ?>>
-								<?php esc_html_e( 'IP Address (backend validation)', 'ffcertificate' ); ?>
-							</label>
-							<p class="description"><?php esc_html_e( 'Choose one or both methods. GPS is more accurate but requires user permission.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Allowed Areas (GPS)', 'ffcertificate' ); ?></label></th>
-						<td>
-							<fieldset>
-								<label>
-									<input type="radio" name="ffc_geofence[geo_area_source]" value="locations" <?php checked( $geo_area_source, 'locations' ); ?>>
-									<?php esc_html_e( 'Registered locations', 'ffcertificate' ); ?>
-								</label>
-								&nbsp;&nbsp;
-								<label>
-									<input type="radio" name="ffc_geofence[geo_area_source]" value="custom" <?php checked( $geo_area_source, 'custom' ); ?>>
-									<?php esc_html_e( 'Custom coordinates', 'ffcertificate' ); ?>
-								</label>
-							</fieldset>
-
-							<div class="<?php echo esc_attr( 'ffc-geo-source-locations' . ( 'locations' !== $geo_area_source ? ' ffc-initially-hidden' : '' ) ); ?>">
-								<select multiple name="ffc_geofence[geo_area_location_ids][]" class="ffc-w100" size="5">
-									<?php foreach ( $all_locations as $loc ) : ?>
-										<option value="<?php echo esc_attr( $loc['id'] ); ?>" <?php echo in_array( $loc['id'], $geo_area_location_ids, true ) ? 'selected' : ''; ?>>
-											<?php echo esc_html( $loc['name'] . ' (' . $loc['lat'] . ', ' . $loc['lng'] . ', ' . $loc['radius'] . 'm)' ); ?>
-										</option>
-									<?php endforeach; ?>
-								</select>
-								<p class="description"><?php esc_html_e( 'Hold Ctrl/Cmd to select multiple locations.', 'ffcertificate' ); ?></p>
-							</div>
-
-							<div class="<?php echo esc_attr( 'ffc-geo-source-custom' . ( 'custom' !== $geo_area_source ? ' ffc-initially-hidden' : '' ) ); ?>">
-								<textarea name="ffc_geofence[geo_areas]" rows="5" class="ffc-w100" placeholder="-23.5505, -46.6333, 5000&#10;-22.9068, -43.1729, 10000"><?php echo esc_textarea( $geo_areas ); ?></textarea>
-								<p class="description"><?php esc_html_e( 'Format: latitude, longitude, radius(meters) - One per line. Example: -23.5505, -46.6333, 5000', 'ffcertificate' ); ?></p>
-							</div>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'IP Geolocation Areas', 'ffcertificate' ); ?></label></th>
-						<td>
-							<label>
-								<input type="checkbox" name="ffc_geofence[geo_ip_areas_permissive]" value="1" <?php checked( $geo_ip_areas_permissive, '1' ); ?>>
-								<?php esc_html_e( 'Use different areas for IP validation', 'ffcertificate' ); ?>
-							</label>
-							<p class="description"><?php esc_html_e( 'When unchecked, IP validation uses the same areas as GPS.', 'ffcertificate' ); ?></p>
-
-							<div class="<?php echo esc_attr( 'ffc-ip-areas-container' . ( '1' !== $geo_ip_areas_permissive ? ' ffc-initially-hidden' : '' ) ); ?>">
-								<br>
-								<fieldset>
-									<label>
-										<input type="radio" name="ffc_geofence[geo_ip_area_source]" value="locations" <?php checked( $geo_ip_area_source, 'locations' ); ?>>
-										<?php esc_html_e( 'Registered locations', 'ffcertificate' ); ?>
-									</label>
-									&nbsp;&nbsp;
-									<label>
-										<input type="radio" name="ffc_geofence[geo_ip_area_source]" value="custom" <?php checked( $geo_ip_area_source, 'custom' ); ?>>
-										<?php esc_html_e( 'Custom coordinates', 'ffcertificate' ); ?>
-									</label>
-								</fieldset>
-
-								<div class="<?php echo esc_attr( 'ffc-geo-source-locations' . ( 'locations' !== $geo_ip_area_source ? ' ffc-initially-hidden' : '' ) ); ?>">
-									<select multiple name="ffc_geofence[geo_ip_area_location_ids][]" class="ffc-w100" size="5">
-										<?php foreach ( $all_locations as $loc ) : ?>
-											<option value="<?php echo esc_attr( $loc['id'] ); ?>" <?php echo in_array( $loc['id'], $geo_ip_area_location_ids, true ) ? 'selected' : ''; ?>>
-												<?php echo esc_html( $loc['name'] . ' (' . $loc['lat'] . ', ' . $loc['lng'] . ', ' . $loc['radius'] . 'm)' ); ?>
-											</option>
-										<?php endforeach; ?>
-									</select>
-									<p class="description"><?php esc_html_e( 'Hold Ctrl/Cmd to select multiple locations.', 'ffcertificate' ); ?></p>
-								</div>
-
-								<div class="<?php echo esc_attr( 'ffc-geo-source-custom' . ( 'custom' !== $geo_ip_area_source ? ' ffc-initially-hidden' : '' ) ); ?>">
-									<textarea name="ffc_geofence[geo_ip_areas]" rows="5" class="ffc-w100" placeholder="-23.5505, -46.6333, 50000&#10;-22.9068, -43.1729, 100000"><?php echo esc_textarea( $geo_ip_areas ); ?></textarea>
-									<p class="description"><?php esc_html_e( 'IP geolocation is less precise (1-50km). Use larger radius (in meters).', 'ffcertificate' ); ?></p>
-								</div>
-							</div>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'GPS + IP Logic', 'ffcertificate' ); ?></label></th>
-						<td>
-							<select name="ffc_geofence[geo_gps_ip_logic]">
-								<option value="or" <?php selected( $geo_gps_ip_logic, 'or' ); ?>><?php esc_html_e( 'OR - Allow if GPS OR IP is valid (recommended)', 'ffcertificate' ); ?></option>
-								<option value="and" <?php selected( $geo_gps_ip_logic, 'and' ); ?>><?php esc_html_e( 'AND - Require both GPS AND IP to be valid (stricter)', 'ffcertificate' ); ?></option>
-							</select>
-							<p class="description"><?php esc_html_e( 'When both GPS and IP are enabled, how to combine the results.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Display Mode', 'ffcertificate' ); ?></label></th>
-						<td>
-							<select name="ffc_geofence[geo_hide_mode]">
-								<option value="message" <?php selected( $geo_hide_mode, 'message' ); ?>><?php esc_html_e( 'Show blocked message (Recommended)', 'ffcertificate' ); ?></option>
-								<option value="title_message" <?php selected( $geo_hide_mode, 'title_message' ); ?>><?php esc_html_e( 'Show title + description + message', 'ffcertificate' ); ?></option>
-								<option value="hide" <?php selected( $geo_hide_mode, 'hide' ); ?>><?php esc_html_e( 'Hide form completely', 'ffcertificate' ); ?></option>
-							</select>
-							<p class="description"><?php esc_html_e( 'How to display the form when user is outside allowed areas.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Blocked Message', 'ffcertificate' ); ?></label></th>
-						<td>
-							<textarea name="ffc_geofence[msg_geo_blocked]" rows="2" class="ffc-w100"><?php echo esc_textarea( $msg_geo_blocked ); ?></textarea>
-							<p class="description"><?php esc_html_e( 'Message shown when user is outside allowed geographic areas.', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-					<tr>
-						<th><label><?php esc_html_e( 'Error Message', 'ffcertificate' ); ?></label></th>
-						<td>
-							<textarea name="ffc_geofence[msg_geo_error]" rows="2" class="ffc-w100"><?php echo esc_textarea( $msg_geo_error ); ?></textarea>
-							<p class="description"><?php esc_html_e( 'Message shown when location detection fails (GPS denied, etc).', 'ffcertificate' ); ?></p>
-						</td>
-					</tr>
-				</table>
-			</div>
-		</div>
-		<script>
-		jQuery(function($) {
-			function toggleGeoSource(prefix) {
-				var source = $('input[name="ffc_geofence[' + prefix + '_source]"]:checked').val();
-				var container = $('input[name="ffc_geofence[' + prefix + '_source]"]').closest('td');
-				container.find('.ffc-geo-source-locations')[source === 'locations' ? 'show' : 'hide']();
-				container.find('.ffc-geo-source-custom')[source === 'custom' ? 'show' : 'hide']();
-			}
-			$('input[name="ffc_geofence[geo_area_source]"]').on('change', function() { toggleGeoSource('geo_area'); });
-			$('input[name="ffc_geofence[geo_ip_area_source]"]').on('change', function() { toggleGeoSource('geo_ip_area'); });
-			toggleGeoSource('geo_area');
-			toggleGeoSource('geo_ip_area');
-
-			$('input[name="ffc_geofence[geo_ip_areas_permissive]"]').on('change', function() {
-				$('.ffc-ip-areas-container')[$(this).is(':checked') ? 'show' : 'hide']();
-			});
-		});
-		</script>
-		<?php
+		$this->geofence->render( $post );
 	}
 
-	/**
-	 * Section 6: Quiz / Evaluation Mode
-	 *
-	 * @param WP_Post $post The post object.
-	 */
 	public function render_box_quiz( WP_Post $post ): void {
-		$config = get_post_meta( $post->ID, '_ffc_form_config', true );
-		if ( ! is_array( $config ) ) {
-			$config = array();
-		}
+		$this->quiz->render( $post );
+	}
 
-		$quiz_enabled       = ( $config['quiz_enabled'] ?? '0' ) === '1';
-		$quiz_passing_score = $config['quiz_passing_score'] ?? '70';
-		$quiz_max_attempts  = $config['quiz_max_attempts'] ?? '0';
-		$quiz_show_score    = ( $config['quiz_show_score'] ?? '1' ) === '1';
-		$quiz_show_correct  = ( $config['quiz_show_correct'] ?? '0' ) === '1';
-		?>
-		<table class="form-table">
-			<tr>
-				<th><label><?php esc_html_e( 'Enable Quiz Mode', 'ffcertificate' ); ?></label></th>
-				<td>
-					<label>
-						<input type="checkbox" name="ffc_config[quiz_enabled]" value="1" id="ffc_quiz_enabled" <?php checked( $quiz_enabled ); ?>>
-						<?php esc_html_e( 'Turn this form into a quiz/evaluation', 'ffcertificate' ); ?>
-					</label>
-					<p class="description"><?php esc_html_e( 'When enabled, radio and select fields can have point values per option. The form is scored on submission.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-			<tr class="ffc-quiz-setting<?php echo $quiz_enabled ? '' : ' ffc-hidden'; ?>">
-				<th><label><?php esc_html_e( 'Passing Score (%)', 'ffcertificate' ); ?></label></th>
-				<td>
-					<input type="number" name="ffc_config[quiz_passing_score]" value="<?php echo esc_attr( $quiz_passing_score ); ?>" min="0" max="100" step="1" class="small-text">
-					<p class="description"><?php esc_html_e( 'Minimum percentage to pass. Set 0 for no minimum.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-			<tr class="ffc-quiz-setting<?php echo $quiz_enabled ? '' : ' ffc-hidden'; ?>">
-				<th><label><?php esc_html_e( 'Max Attempts', 'ffcertificate' ); ?></label></th>
-				<td>
-					<input type="number" name="ffc_config[quiz_max_attempts]" value="<?php echo esc_attr( $quiz_max_attempts ); ?>" min="0" step="1" class="small-text">
-					<p class="description"><?php esc_html_e( 'Max retries per CPF/RF. 0 = unlimited.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-			<tr class="ffc-quiz-setting<?php echo $quiz_enabled ? '' : ' ffc-hidden'; ?>">
-				<th><label><?php esc_html_e( 'Display Options', 'ffcertificate' ); ?></label></th>
-				<td>
-					<label>
-						<input type="checkbox" name="ffc_config[quiz_show_score]" value="1" <?php checked( $quiz_show_score ); ?>>
-						<?php esc_html_e( 'Show score after submission', 'ffcertificate' ); ?>
-					</label><br>
-					<label>
-						<input type="checkbox" name="ffc_config[quiz_show_correct]" value="1" <?php checked( $quiz_show_correct ); ?>>
-						<?php esc_html_e( 'Show which answers were correct/incorrect', 'ffcertificate' ); ?>
-					</label>
-				</td>
-			</tr>
-		</table>
-		<?php
-		// Quiz toggle logic in ffc-admin.js.
+	public function render_box_public_csv_download( WP_Post $post ): void {
+		$this->public_csv_download->render( $post );
+	}
+
+	public function render_box_device_limit( WP_Post $post ): void {
+		$this->device_limit->render( $post );
 	}
 
 	/**
-	 * Helper: Renders a field row in the builder
-	 *
 	 * @param int|string           $index Field index.
 	 * @param array<string, mixed> $field Field data.
 	 */
 	public function render_field_row( $index, array $field ): void {
-		$index     = (string) $index;
-		$type      = isset( $field['type'] ) ? $field['type'] : 'text';
-		$label     = isset( $field['label'] ) ? $field['label'] : '';
-		$name      = isset( $field['name'] ) ? $field['name'] : '';
-		$req       = isset( $field['required'] ) ? $field['required'] : '';
-		$opts      = isset( $field['options'] ) ? $field['options'] : '';
-		$content   = isset( $field['content'] ) ? $field['content'] : '';
-		$embed_url = isset( $field['embed_url'] ) ? $field['embed_url'] : '';
-		$points    = isset( $field['points'] ) ? $field['points'] : '';
-
-		$is_info               = 'info' === $type;
-		$is_embed              = 'embed' === $type;
-		$is_display_only       = $is_info || $is_embed;
-		$options_visible_class = ( 'select' === $type || 'radio' === $type ) ? '' : 'ffc-hidden';
-		?>
-		<div class="ffc-field-row" data-index="<?php echo esc_attr( $index ); ?>">
-			<div class="ffc-field-row-header">
-				<span class="ffc-sort-handle">
-					<span class="dashicons dashicons-menu"></span>
-					<span class="ffc-field-title"><strong>
-					<?php
-					if ( $is_info ) {
-						esc_html_e( 'Info Block', 'ffcertificate' );
-					} elseif ( $is_embed ) {
-						esc_html_e( 'Embed', 'ffcertificate' );
-					} else {
-						esc_html_e( 'Field', 'ffcertificate' );
-					}
-					?>
-					</strong></span>
-				</span>
-				<button type="button" class="button button-link-delete ffc-remove-field"><?php esc_html_e( 'Remove', 'ffcertificate' ); ?></button>
-			</div>
-
-			<div class="ffc-field-row-grid">
-				<div class="ffc-grid-item">
-					<label>
-					<?php
-					if ( $is_info ) {
-						esc_html_e( 'Title (optional)', 'ffcertificate' );
-					} elseif ( $is_embed ) {
-						esc_html_e( 'Caption (optional)', 'ffcertificate' );
-					} else {
-						esc_html_e( 'Label', 'ffcertificate' );
-					}
-					?>
-					</label>
-					<input type="text" name="ffc_fields[<?php echo esc_attr( $index ); ?>][label]" value="<?php echo esc_attr( $label ); ?>" class="ffc-w100">
-				</div>
-				<div class="ffc-grid-item ffc-standard-row<?php echo $is_display_only ? ' ffc-hidden' : ''; ?>">
-					<label><?php esc_html_e( 'Variable Name (Tag)', 'ffcertificate' ); ?></label>
-					<input type="text" name="ffc_fields[<?php echo esc_attr( $index ); ?>][name]" value="<?php echo esc_attr( $name ); ?>" placeholder="<?php esc_attr_e( 'ex: course_name', 'ffcertificate' ); ?>" class="ffc-w100">
-				</div>
-				<div class="ffc-grid-item">
-					<label><?php esc_html_e( 'Type', 'ffcertificate' ); ?></label>
-					<select name="ffc_fields[<?php echo esc_attr( $index ); ?>][type]" class="ffc-field-type-selector ffc-w100">
-						<option value="text" <?php selected( $type, 'text' ); ?>><?php esc_html_e( 'Text', 'ffcertificate' ); ?></option>
-						<option value="email" <?php selected( $type, 'email' ); ?>><?php esc_html_e( 'Email', 'ffcertificate' ); ?></option>
-						<option value="number" <?php selected( $type, 'number' ); ?>><?php esc_html_e( 'Number', 'ffcertificate' ); ?></option>
-						<option value="date" <?php selected( $type, 'date' ); ?>><?php esc_html_e( 'Date', 'ffcertificate' ); ?></option>
-						<option value="textarea" <?php selected( $type, 'textarea' ); ?>><?php esc_html_e( 'Textarea', 'ffcertificate' ); ?></option>
-						<option value="select" <?php selected( $type, 'select' ); ?>><?php esc_html_e( 'Select (Combobox)', 'ffcertificate' ); ?></option>
-						<option value="radio" <?php selected( $type, 'radio' ); ?>><?php esc_html_e( 'Radio Box', 'ffcertificate' ); ?></option>
-						<option value="hidden" <?php selected( $type, 'hidden' ); ?>><?php esc_html_e( 'Hidden Field', 'ffcertificate' ); ?></option>
-						<option value="info" <?php selected( $type, 'info' ); ?>><?php esc_html_e( 'Info Block', 'ffcertificate' ); ?></option>
-						<option value="embed" <?php selected( $type, 'embed' ); ?>><?php esc_html_e( 'Embed (Media)', 'ffcertificate' ); ?></option>
-					</select>
-				</div>
-				<div class="ffc-grid-item ffc-flex-center ffc-standard-row<?php echo $is_display_only ? ' ffc-hidden' : ''; ?>">
-					<label class="ffc-req-label">
-						<input type="checkbox" name="ffc_fields[<?php echo esc_attr( $index ); ?>][required]" value="1" <?php checked( $req, '1' ); ?>>
-						<?php esc_html_e( 'Required?', 'ffcertificate' ); ?>
-					</label>
-				</div>
-			</div>
-
-			<div class="ffc-content-field<?php echo $is_info ? '' : ' ffc-hidden'; ?>">
-				<p class="description ffc-options-desc">
-					<?php esc_html_e( 'Content (supports <b>, <i>, <a>, <p>, <ul>, <ol>):', 'ffcertificate' ); ?>
-				</p>
-				<textarea name="ffc_fields[<?php echo esc_attr( $index ); ?>][content]" rows="4" class="ffc-w100"><?php echo esc_textarea( $content ); ?></textarea>
-			</div>
-
-			<div class="ffc-embed-field<?php echo $is_embed ? '' : ' ffc-hidden'; ?>">
-				<p class="description ffc-options-desc">
-					<?php esc_html_e( 'Media URL (YouTube, Vimeo, image, or audio):', 'ffcertificate' ); ?>
-				</p>
-				<input type="url" name="ffc_fields[<?php echo esc_attr( $index ); ?>][embed_url]" value="<?php echo esc_url( $embed_url ); ?>" placeholder="https://www.youtube.com/watch?v=..." class="ffc-w100">
-			</div>
-
-			<div class="ffc-options-field <?php echo esc_attr( $options_visible_class ); ?>">
-				<p class="description ffc-options-desc">
-					<?php esc_html_e( 'Options (separate with commas):', 'ffcertificate' ); ?>
-				</p>
-				<input type="text" name="ffc_fields[<?php echo esc_attr( $index ); ?>][options]" value="<?php echo esc_attr( $opts ); ?>" placeholder="<?php esc_attr_e( 'Ex: Option 1, Option 2, Option 3', 'ffcertificate' ); ?>" class="ffc-w100">
-				<div class="ffc-quiz-points ffc-hidden">
-					<p class="description ffc-options-desc ffc-mt5">
-						<?php esc_html_e( 'Points per option (same order, separate with commas):', 'ffcertificate' ); ?>
-					</p>
-					<input type="text" name="ffc_fields[<?php echo esc_attr( $index ); ?>][points]" value="<?php echo esc_attr( $points ); ?>" placeholder="<?php esc_attr_e( 'Ex: 0, 10, 0', 'ffcertificate' ); ?>" class="ffc-w100">
-				</div>
-			</div>
-		</div>
-		<?php
-	}
-
-	/**
-	 * Section 7: Public CSV Download (shortcode-powered frontend export).
-	 *
-	 * Lets the admin enable a public download link guarded by a revocable
-	 * hash + usage quota. The actual download endpoint is handled by
-	 * {@see \FreeFormCertificate\Frontend\PublicCsvDownload}.
-	 *
-	 * @since 5.1.0
-	 * @param WP_Post $post The post object.
-	 */
-	public function render_box_public_csv_download( WP_Post $post ): void {
-		$enabled = (string) get_post_meta( $post->ID, '_ffc_csv_public_enabled', true );
-		$hash    = (string) get_post_meta( $post->ID, '_ffc_csv_public_hash', true );
-		$limit   = (int) get_post_meta( $post->ID, '_ffc_csv_public_limit', true );
-		$count   = (int) get_post_meta( $post->ID, '_ffc_csv_public_count', true );
-
-		if ( $limit <= 0 ) {
-			$settings      = get_option( 'ffc_settings', array() );
-			$default_limit = ( is_array( $settings ) && isset( $settings['public_csv_default_limit'] ) )
-				? (int) $settings['public_csv_default_limit']
-				: 1;
-			$limit         = $default_limit > 0 ? $default_limit : 1;
-		}
-
-		// Check that a geofence end date is configured — required for this feature.
-		$geofence_config = get_post_meta( $post->ID, '_ffc_geofence_config', true );
-		$date_end        = ( is_array( $geofence_config ) && ! empty( $geofence_config['date_end'] ) )
-			? (string) $geofence_config['date_end']
-			: '';
-
-		// Sub-fields are disabled (read-only-ish) when the master toggle is off.
-		// JS mirrors this on change so the UI updates without saving. The save
-		// handler also short-circuits when enabled=0 to preserve persisted
-		// values that the disabled inputs would not have submitted.
-		$sub_disabled = ( '1' !== $enabled );
-
-		// Nonce is emitted by render_box_layout(), which always renders before this metabox.
-		?>
-		<p class="description">
-			<?php esc_html_e( 'Allow a person without WordPress access to download the submissions CSV for this form by providing the Form ID + a hash on a public page that contains the [ffc_csv_download] shortcode.', 'ffcertificate' ); ?>
-		</p>
-
-		<input type="hidden" name="ffc_csv_public[present]" value="1">
-		<table class="form-table ffc-csv-public-table<?php echo $sub_disabled ? ' ffc-csv-public-disabled' : ''; ?>">
-			<tr>
-				<th scope="row">
-					<label for="ffc_csv_public_enabled">
-						<?php esc_html_e( 'Enable Public Download', 'ffcertificate' ); ?>
-					</label>
-				</th>
-				<td>
-					<label>
-						<input type="checkbox"
-								name="ffc_csv_public[enabled]"
-								id="ffc_csv_public_enabled"
-								value="1"
-								<?php checked( $enabled, '1' ); ?>>
-						<?php esc_html_e( 'Allow visitors with the hash to download this form\'s CSV.', 'ffcertificate' ); ?>
-					</label>
-
-					<?php if ( '' === $date_end ) : ?>
-						<p class="description ffc-warning-text">
-							<strong><?php esc_html_e( 'Warning:', 'ffcertificate' ); ?></strong>
-							<?php esc_html_e( 'This form has no end date configured in the "Geolocation & Date/Time Restrictions" metabox. You must set an end date before public downloads will work — downloads are only released after the form has ended.', 'ffcertificate' ); ?>
-						</p>
-					<?php endif; ?>
-				</td>
-			</tr>
-
-			<tr class="ffc-csv-public-sub">
-				<th scope="row">
-					<label for="ffc_csv_public_limit">
-						<?php esc_html_e( 'Download Limit', 'ffcertificate' ); ?>
-					</label>
-				</th>
-				<td>
-					<input type="number"
-							name="ffc_csv_public[limit]"
-							id="ffc_csv_public_limit"
-							min="1"
-							step="1"
-							class="small-text"
-							value="<?php echo esc_attr( (string) $limit ); ?>"<?php disabled( $sub_disabled ); ?>>
-					<p class="description">
-						<?php
-						printf(
-							/* translators: 1: current count, 2: limit */
-							esc_html__( 'Maximum number of CSV downloads allowed via the public page. Current usage: %1$d of %2$d.', 'ffcertificate' ),
-							(int) $count,
-							(int) $limit
-						);
-						?>
-					</p>
-				</td>
-			</tr>
-
-			<tr class="ffc-csv-public-sub">
-				<th scope="row">
-					<label for="ffc_csv_public_hash">
-						<?php esc_html_e( 'Access Hash', 'ffcertificate' ); ?>
-					</label>
-				</th>
-				<td>
-					<input type="text"
-							id="ffc_csv_public_hash"
-							value="<?php echo esc_attr( $hash ); ?>"
-							readonly
-							class="large-text code"
-							onclick="this.select();">
-					<p class="ffc-mt-10">
-						<label>
-							<input type="checkbox" name="ffc_csv_public[regenerate_hash]" value="1"<?php disabled( $sub_disabled ); ?>>
-							<?php esc_html_e( 'Regenerate hash on save (invalidates the current link).', 'ffcertificate' ); ?>
-						</label>
-					</p>
-					<p>
-						<label>
-							<input type="checkbox" name="ffc_csv_public[reset_counter]" value="1"<?php disabled( $sub_disabled ); ?>>
-							<?php esc_html_e( 'Reset the download counter to zero on save.', 'ffcertificate' ); ?>
-						</label>
-					</p>
-					<p class="description">
-						<?php esc_html_e( 'Share the Form ID and this hash with the person who should download the CSV. If no hash exists yet, one will be generated automatically when you enable the feature and save.', 'ffcertificate' ); ?>
-					</p>
-
-					<?php if ( '1' === $enabled && '' !== $hash ) : ?>
-						<?php
-						$ffc_settings          = get_option( 'ffc_settings', array() );
-						$csv_download_page_url = $ffc_settings['csv_download_page_url'] ?? '';
-						$ffc_query_string      = 'form_id=' . $post->ID . '&hash=' . $hash;
-						?>
-						<?php if ( '' !== $csv_download_page_url ) : ?>
-							<p class="description">
-								<?php esc_html_e( 'Download link:', 'ffcertificate' ); ?>
-								<br>
-								<code><?php echo esc_html( $csv_download_page_url . ( str_contains( $csv_download_page_url, '?' ) ? '&' : '?' ) . $ffc_query_string ); ?></code>
-							</p>
-						<?php else : ?>
-							<p class="description">
-								<?php esc_html_e( 'Example pre-filled link (append this as a query string to the page that contains the [ffc_csv_download] shortcode):', 'ffcertificate' ); ?>
-								<br>
-								<code>?<?php echo esc_html( $ffc_query_string ); ?></code>
-							</p>
-						<?php endif; ?>
-					<?php endif; ?>
-				</td>
-			</tr>
-
-			<?php
-			$cpf_mode      = (string) get_post_meta( $post->ID, '_ffc_csv_public_cpf_mode', true );
-			$cpf_whitelist = (string) get_post_meta( $post->ID, '_ffc_csv_public_cpf_whitelist', true );
-			if ( '' === $cpf_mode ) {
-				$cpf_mode = 'none';
-			}
-			?>
-			<tr class="ffc-csv-public-sub">
-				<th scope="row">
-					<label for="ffc_csv_public_cpf_mode"><?php esc_html_e( 'Require CPF for download', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<select name="ffc_csv_public[cpf_mode]" id="ffc_csv_public_cpf_mode"<?php disabled( $sub_disabled ); ?>>
-						<option value="none" <?php selected( $cpf_mode, 'none' ); ?>><?php esc_html_e( 'No — only Form ID + Hash', 'ffcertificate' ); ?></option>
-						<option value="audit" <?php selected( $cpf_mode, 'audit' ); ?>><?php esc_html_e( 'Audit — require CPF, but do not match against any list', 'ffcertificate' ); ?></option>
-						<option value="participants" <?php selected( $cpf_mode, 'participants' ); ?>><?php esc_html_e( 'Participants — CPF must match a submission', 'ffcertificate' ); ?></option>
-						<option value="owner" <?php selected( $cpf_mode, 'owner' ); ?>><?php esc_html_e( 'Owner — CPF must match the form author', 'ffcertificate' ); ?></option>
-						<option value="whitelist" <?php selected( $cpf_mode, 'whitelist' ); ?>><?php esc_html_e( 'Whitelist — CPF must be in the list below', 'ffcertificate' ); ?></option>
-					</select>
-					<p class="description">
-						<?php esc_html_e( 'When set to anything other than "No", every download attempt (with success/failure flag and the CPF encrypted at rest) is recorded in an audit log on this form, capped at the most recent 100 attempts. Use the "Download audit log (CSV)" button below to export.', 'ffcertificate' ); ?>
-					</p>
-				</td>
-			</tr>
-
-			<?php
-			// Whitelist textarea is rendered only when the persisted mode is
-			// 'whitelist'. Switching the dropdown alone does not reveal it —
-			// the user must save the form first. This keeps the textarea
-			// invisible while the mode is in flux and matches the prompt
-			// "só deve aparecer se for selecionada essa opção e salvar".
-			if ( 'whitelist' === $cpf_mode ) :
-				?>
-			<tr class="ffc-csv-public-sub">
-				<th scope="row">
-					<label for="ffc_csv_public_cpf_whitelist"><?php esc_html_e( 'CPF whitelist', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<textarea name="ffc_csv_public[cpf_whitelist]"
-						id="ffc_csv_public_cpf_whitelist"
-						rows="4"
-						class="large-text code"
-						placeholder="000.000.000-00&#10;111.111.111-11"<?php disabled( $sub_disabled ); ?>><?php echo esc_textarea( $cpf_whitelist ); ?></textarea>
-					<p class="description">
-						<?php esc_html_e( 'One CPF per line. Only used when the mode above is set to "Whitelist". Formatting is ignored — only digits matter.', 'ffcertificate' ); ?>
-					</p>
-				</td>
-			</tr>
-			<?php elseif ( '1' === $enabled ) : ?>
-			<tr class="ffc-csv-public-sub">
-				<th scope="row"></th>
-				<td>
-					<p class="description">
-						<em><?php esc_html_e( 'Tip: select "Whitelist" above and save the form to reveal the CPF whitelist textarea.', 'ffcertificate' ); ?></em>
-					</p>
-				</td>
-			</tr>
-			<?php endif; ?>
-
-			<?php
-			$ffc_audit_summary = \FreeFormCertificate\Frontend\PublicCsvDownload::get_audit_log_summary( $post->ID );
-			?>
-			<tr class="ffc-csv-public-sub">
-				<th scope="row">
-					<?php esc_html_e( 'Download audit log', 'ffcertificate' ); ?>
-				</th>
-				<td>
-					<?php if ( $ffc_audit_summary['count'] > 0 ) : ?>
-						<div class="ffc-csv-audit-summary" role="group" aria-label="<?php esc_attr_e( 'Download attempts breakdown', 'ffcertificate' ); ?>">
-							<div class="ffc-csv-audit-card">
-								<span class="ffc-csv-audit-card-label"><?php esc_html_e( 'Total attempts', 'ffcertificate' ); ?></span>
-								<span class="ffc-csv-audit-card-value"><?php echo esc_html( (string) $ffc_audit_summary['count'] ); ?></span>
-							</div>
-							<div class="ffc-csv-audit-card is-success">
-								<span class="ffc-csv-audit-card-label"><?php esc_html_e( 'Successful', 'ffcertificate' ); ?></span>
-								<span class="ffc-csv-audit-card-value"><?php echo esc_html( (string) $ffc_audit_summary['success'] ); ?></span>
-							</div>
-							<div class="ffc-csv-audit-card is-fail">
-								<span class="ffc-csv-audit-card-label"><?php esc_html_e( 'Failed', 'ffcertificate' ); ?></span>
-								<span class="ffc-csv-audit-card-value"><?php echo esc_html( (string) $ffc_audit_summary['fail'] ); ?></span>
-							</div>
-						</div>
-						<?php if ( $ffc_audit_summary['count'] >= \FreeFormCertificate\Frontend\PublicCsvDownload::DOWNLOAD_LOG_MAX ) : ?>
-							<p class="description">
-								<?php
-								echo esc_html(
-									sprintf(
-										/* translators: %d: maximum entries kept in the log */
-										__( 'Limit of %d entries reached — older entries roll off.', 'ffcertificate' ),
-										\FreeFormCertificate\Frontend\PublicCsvDownload::DOWNLOAD_LOG_MAX
-									)
-								);
-								?>
-							</p>
-						<?php endif; ?>
-					<?php endif; ?>
-					<?php if ( null !== $ffc_audit_summary['url'] ) : ?>
-						<p>
-							<a href="<?php echo esc_url( $ffc_audit_summary['url'] ); ?>"
-								class="button"
-								download>
-								<?php esc_html_e( 'Download audit log (CSV)', 'ffcertificate' ); ?>
-							</a>
-						</p>
-						<p class="description">
-							<?php
-							if ( class_exists( '\FreeFormCertificate\Core\Encryption' )
-								&& \FreeFormCertificate\Core\Encryption::is_configured() ) {
-								esc_html_e( 'CPFs are stored encrypted at-rest and decrypted on the fly when you download the CSV. Each entry includes timestamp, IP, gate mode, CPF, and outcome.', 'ffcertificate' );
-							} else {
-								esc_html_e( 'Encryption is not configured on this site. The CPF column will be empty for entries written while encryption is off.', 'ffcertificate' );
-							}
-							?>
-						</p>
-					<?php else : ?>
-						<p class="description">
-							<?php esc_html_e( 'No download attempts have been logged yet.', 'ffcertificate' ); ?>
-						</p>
-					<?php endif; ?>
-				</td>
-			</tr>
-		</table>
-		<?php
-	}
-
-	/**
-	 * Render Device Fingerprint Limit metabox.
-	 *
-	 * Per-form override layer for the global Rate Limit → Device Fingerprint
-	 * settings. Empty fields fall through to the global defaults; the
-	 * "Enable for this form" checkbox controls whether the limit applies
-	 * to this form at all.
-	 *
-	 * @param WP_Post $post Form post.
-	 */
-	public function render_box_device_limit( WP_Post $post ): void {
-		$enabled       = (string) get_post_meta( $post->ID, '_ffc_device_limit_enabled', true );
-		$max           = (string) get_post_meta( $post->ID, '_ffc_device_limit_max', true );
-		$threshold     = (string) get_post_meta( $post->ID, '_ffc_device_match_threshold', true );
-		$message       = (string) get_post_meta( $post->ID, '_ffc_device_limit_message', true );
-		$global_active = false;
-		if ( class_exists( '\FreeFormCertificate\Security\RateLimiter' ) ) {
-			$rl_settings   = \FreeFormCertificate\Security\RateLimiter::get_settings();
-			$global_active = ! empty( $rl_settings['device']['enabled'] );
-		}
-
-		// Two independent gates control whether the sub-fields can be edited:
-		// 1. The global Device Fingerprint subsystem (Settings → Rate Limit).
-		// When that is OFF, every input including the master checkbox is
-		// locked — the warning explains how to enable it.
-		// 2. The per-form "Enable for this form" checkbox. When that is OFF
-		// (but global is ON), only the secondary fields lock; the
-		// checkbox itself stays editable so the user can flip it on.
-		$sub_disabled = ( ! $global_active ) || ( '1' !== $enabled );
-
-		$table_classes = array( 'form-table', 'ffc-device-limit-table' );
-		if ( ! $global_active ) {
-			$table_classes[] = 'ffc-device-limit-globally-off';
-		} elseif ( '1' !== $enabled ) {
-			$table_classes[] = 'ffc-device-limit-disabled';
-		}
-
-		// Nonce is emitted by render_box_layout(), which always renders before this metabox.
-		?>
-		<p class="description">
-			<?php esc_html_e( 'Limit how many times the same physical device can submit this form. Combines a persistent cookie with a multi-signal browser fingerprint and the global "N of M" matching rule.', 'ffcertificate' ); ?>
-		</p>
-		<?php if ( ! $global_active ) : ?>
-			<p class="description ffc-warning-text">
-				<strong><?php esc_html_e( 'Note:', 'ffcertificate' ); ?></strong>
-				<?php
-				echo wp_kses(
-					sprintf(
-						/* translators: %s: link to global rate-limit settings */
-						__( 'The Device Fingerprint subsystem is disabled globally. Enable it in %s before configuring per-form overrides.', 'ffcertificate' ),
-						'<a href="' . esc_url( admin_url( 'edit.php?post_type=ffc_form&page=ffc-settings&tab=rate_limit' ) ) . '">' . esc_html__( 'Settings → Rate Limit → Device Fingerprint', 'ffcertificate' ) . '</a>'
-					),
-					array( 'a' => array( 'href' => array() ) )
-				);
-				?>
-			</p>
-		<?php endif; ?>
-
-		<input type="hidden" name="ffc_device_limit[present]" value="1">
-		<table class="<?php echo esc_attr( implode( ' ', $table_classes ) ); ?>">
-			<tr>
-				<th scope="row">
-					<label for="ffc_device_limit_enabled"><?php esc_html_e( 'Enable for this form', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<label>
-						<input type="checkbox"
-							name="ffc_device_limit[enabled]"
-							id="ffc_device_limit_enabled"
-							value="1"
-							<?php checked( $enabled, '1' ); ?><?php disabled( ! $global_active ); ?>>
-						<?php esc_html_e( 'Apply the device-fingerprint limit to this form.', 'ffcertificate' ); ?>
-					</label>
-				</td>
-			</tr>
-
-			<tr class="ffc-device-limit-sub">
-				<th scope="row">
-					<label for="ffc_device_limit_max"><?php esc_html_e( 'Max submissions per device', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<input type="number"
-						name="ffc_device_limit[max]"
-						id="ffc_device_limit_max"
-						min="1"
-						max="100"
-						value="<?php echo esc_attr( $max ); ?>"
-						placeholder="<?php esc_attr_e( 'Default: 2', 'ffcertificate' ); ?>"<?php disabled( $sub_disabled ); ?>>
-					<p class="description"><?php esc_html_e( 'Defaults to 2 when this metabox is enabled. Override here to set a per-form value.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-
-			<tr class="ffc-device-limit-sub">
-				<th scope="row">
-					<label for="ffc_device_limit_threshold"><?php esc_html_e( 'Match threshold (3-12)', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<input type="number"
-						name="ffc_device_limit[threshold]"
-						id="ffc_device_limit_threshold"
-						min="3"
-						max="12"
-						value="<?php echo esc_attr( $threshold ); ?>"
-						placeholder="<?php esc_attr_e( 'Inherit from global', 'ffcertificate' ); ?>"<?php disabled( $sub_disabled ); ?>>
-					<p class="description"><?php esc_html_e( 'Lower = more aggressive. Leave empty to inherit the global default.', 'ffcertificate' ); ?></p>
-				</td>
-			</tr>
-
-			<tr class="ffc-device-limit-sub">
-				<th scope="row">
-					<label for="ffc_device_limit_message"><?php esc_html_e( 'Block message', 'ffcertificate' ); ?></label>
-				</th>
-				<td>
-					<textarea name="ffc_device_limit[message]"
-						id="ffc_device_limit_message"
-						rows="2"
-						class="large-text"
-						placeholder="<?php esc_attr_e( 'Inherit from global', 'ffcertificate' ); ?>"<?php disabled( $sub_disabled ); ?>><?php echo esc_textarea( $message ); ?></textarea>
-				</td>
-			</tr>
-		</table>
-		<?php
+		$this->builder->render_field_row( $index, $field );
 	}
 }
